% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_softmax.R
\name{get_softmax}
\alias{get_softmax}
\title{A Softmax Function}
\usage{
get_softmax(normalized.counts, parameters)
}
\arguments{
\item{normalized.counts}{Normalized counts of cells.}

\item{parameters}{The weights of the neural network.}
}
\description{
This function generates the softmax output for the neural network.
}
\examples{
## An example using human pancreatic GERAS.
## Model file: GERAS_Tf_Hs_Final.rds. 
##Read counts: H1_pRPM.csv 

human_model <- readRDS(file = "GERAS_Tf_Hs_Final.rds")
H1 <- read.table(file = "H1_pRPM.csv", sep = ",", header = T, row.names = 1, check.names = F)

get_softmax(H1.log, human_model$weight)
}
\keyword{GERAS}
